expense-tracker-api/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
PyJWT==4.2.0